Pi-hole - Init
Sets up Pi-hole with the web UI on an Ingress and DNS on a LoadBalancer
IP. The latter is done so Pi-hole can see the source of a query.

Additionally, because the nodes rely on it for DNS, the image needs to
be pre-pulled so it's already available when it restarts. Adds a
DaemonSet to do that, which needs to be updated prior to updating the
actual deployment.