Upgrade to Loki 2.9.4 and k8s lib 1.28
Loki and Promtail environments both upgraded to 2.9.4 via jb update,
then apply. Minor tweak needed to the config due to a new storage class
variable.

Also updates k8s API libsonnet to 1.28, since the cluster is on that
now. Learned the lesson of putting the version number path in every
single file, instead of in a central library file. Fixed that.