Migrate cert-manager to HelmChart resource
K3s has a helm-controller built-in, so I'm just using that:
https://docs.k3s.io/helm

Created a quick helper library for creating that resource, and migrated
cert-manager to it as a simple test.

HelmChart resources are created in a separate dedicated namespace
because a Job gets spun up for each Helm operation. Makes it easier to
query/monitor if they're all in one place, plus provides easier
isolation if needed.