Override Loki container memory requests
The defaults on these are tuned for a large deployment ingesting
terabytes of logs per day. Needless to say, I'm not doing that. I just
wanted high availability (and I like that it's jsonnet libraries vs
Helm).

Reducing because I don't have that much memory in my cluster.
For reference, these values are less than 10% of the defaults.