Add monitoring dashboards for Loki
Uses the loki-mixin, but it was designed for use with kube-prometheus,
so I had to pick out the things I wanted individually.

Also had to modify a few things to get some of the metrics aligned with
what the dashboards expect. There's still a few that don't work,
possibly because kube-state-metrics is missing. I'll get to that.