Add monitoring for Mimir
Adds dashboards for Grafana and ServiceMonitor + recording rules for
Mimir.

Mimir's docs are worded a little differently, and made it more clear
that the dashboards they've designed are using the "cluster" label to
identify a hosting k8s cluster/equivalent, NOT just the Prometheus
cluster. This helps me understand some of the decisions they made on the
Loki dashboards, but means I'll have to tweak a few more things.