Update MinIO, set FSGroupChangePolicy to OnRootMismatch
The latter is what was causing my MinIO pods to take 8+ minutes to
restart. The Operator will default to this, but only if you haven't set
it before. Somewhere along the line, this was populated by a default
when exported to a file, so it stuck that way.

I really don't like how MinIO's Operator works. I can't wait to put this
in jsonnet; I will remove any fields that are already the "default"
value at that time so shit like this doesn't keep happening.