Libs rewrite
I started rewriting some of the helper functions to try to make it
easier to deploy things to a test cluster, then got fed up with the
amount of arguments I was having to pass around.

Re-wrote many of the "standard" resource functions to accept a config
object instead. Also generalized most of the functions in the
weebcluster helper and moved them to utils for consistency.